requirements.txt
openai
streamlit
sounddevice
numpy
python-dotenv
openai-whisper